working tsp comparison with uct, ments, mmcts

+++ mcts/sampling/uct_distribution.py
@@ -22,7 +22,7 @@ def uct_distribution(puct_values: torch.Tensor, inf_value_cap_coeff: float = 10.
     # softmax
     puct_probabilities = torch.softmax(capped_puct, dim=0)
 
-    return puct_probabilities
+    return puct_probabilities.to(dtype=torch.float32)
 
 
 if __name__ == "__main__":
